fix: fix some code and features error

+++ src/main/java/com/application/bookstore/service/book/BookServiceImpl.java
@@ -34,6 +34,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteBookById(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find book by id: " + id);
+        }
         bookRepository.deleteById(id);
     }
